reorganise and add tests
