+++ test/solidity/contracts/addressLocal.js
@@ -1,0 +1,10 @@
+'use strict'
+
+module.exports = {
+  contract: `
+contract addressLocal {
+      function addressLocal () {
+          address sender = msg.sender;
+      }
+  }
+`}
